refactor: move SourcesTable into dedicated file in ui folder
