+++ src/components/ui/Tables/DestinationsTable/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./DestinationsTable";
+export * from "./DestinationsTable";
