+++ src/components/ui/Tables/SourcesTable/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./SourcesTable";
+export * from "./SourcesTable";
